+++ thoth-export-server/src/marc21/marc21record_thoth.rs
@@ -634,8 +634,8 @@ impl Marc21Field<Marc21RecordThoth> for WorkIssues {
                 .and_then(|f| {
                     f.add_subfield(b"v", format!("vol. {}.", self.issue_ordinal).as_bytes())
                 })
-                .and_then(|f| f.add_subfield(b"x", self.series.issn_digital.as_bytes()))
-                .and_then(|f| f.add_subfield(b"x", self.series.issn_print.as_bytes()))
+                .and_then(|f| f.add_subfield(b"x", self.series.issn_digital.as_deref().unwrap_or_default().as_bytes()))
+                .and_then(|f| f.add_subfield(b"x", self.series.issn_print.as_deref().unwrap_or_default().as_bytes()))
                 .and_then(|f| builder.add_field(f))?;
         }
         Ok(())
@@ -832,8 +832,8 @@ pub(crate) mod tests {
                 series: WorkIssuesSeries {
                     series_type: SeriesType::BOOK_SERIES,
                     series_name: "Name of series".to_string(),
-                    issn_print: "1234-5678".to_string(),
-                    issn_digital: "8765-4321".to_string(),
+                    issn_print: Some("1234-5678".to_string()),
+                    issn_digital: Some("8765-4321".to_string()),
                     series_url: None,
                     series_description: None,
                     series_cfp_url: None,