+++ thoth-export-server/src/xml/onix21_proquest_ebrary.rs
@@ -766,8 +766,8 @@ mod tests {
             series: WorkIssuesSeries {
                 series_type: thoth_client::SeriesType::JOURNAL,
                 series_name: "Name of series".to_string(),
-                issn_print: "1234-5678".to_string(),
-                issn_digital: "8765-4321".to_string(),
+                issn_print: Some("1234-5678".to_string()),
+                issn_digital: Some("8765-4321".to_string()),
                 series_url: None,
                 series_description: None,
                 series_cfp_url: None,