+++ thoth-export-server/src/xml/onix3_google_books.rs
@@ -609,7 +609,7 @@ impl XmlElementBlock<Onix3GoogleBooks> for WorkIssues {
                 })?;
                 write_element_block("IDValue", w, |w| {
                     w.write(XmlEvent::Characters(
-                        &self.series.issn_digital.replace('-', ""),
+                        &self.series.issn_digital.as_deref().unwrap_or_default().replace('-', ""),
                     ))
                     .map_err(|e| e.into())
                 })
@@ -778,8 +778,8 @@ mod tests {
             series: WorkIssuesSeries {
                 series_type: thoth_client::SeriesType::JOURNAL,
                 series_name: "Name of series".to_string(),
-                issn_print: "1234-5678".to_string(),
-                issn_digital: "8765-4321".to_string(),
+                issn_print: Some("1234-5678".to_string()),
+                issn_digital: Some("8765-4321".to_string()),
                 series_url: None,
                 series_description: None,
                 series_cfp_url: None,
@@ -803,7 +803,7 @@ mod tests {
         // Change all possible values to test that output is updated
         test_issue.issue_ordinal = 2;
         test_issue.series.series_name = "Different series".to_string();
-        test_issue.series.issn_digital = "1111-2222".to_string();
+        test_issue.series.issn_digital = Some("1111-2222".to_string());
         let output = generate_test_output(true, &test_issue);
         assert!(output.contains(r#"<Collection>"#));
         assert!(output.contains(r#"  <CollectionType>10</CollectionType>"#));