+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - [475](https://github.com/thoth-pub/thoth/issues/475) - Add seconds to timestamp for Crossref metadata output
   - [571](https://github.com/thoth-pub/thoth/issues/571) - Fix overlapping URL text for Locations in Thoth Admin panel on website in Safari and Chromium browsers
 
+### Changed
+  - [218](https://github.com/thoth-pub/thoth/issues/218) - Make series ISSN optional
+
 ### Security
   - [572](https://github.com/thoth-pub/thoth/pull/572) - Upgrade `mio` to v0.8.11